one_stop/requirements.txt
asgiref==3.5.0
asttokens==2.0.5
autopep8==1.6.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.11.1
black==22.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
crispy-bootstrap5==0.6
cryptography==36.0.1
cx-Oracle==8.3.0
debugpy==1.5.1
decorator==5.1.1
dj-database-url==0.5.0
Django==4.0.6
django-appconf==1.0.5
django-auth-ldap==4.0.0
django-bootstrap5==21.3
django-crispy-forms==1.14.0
django-imagekit==4.1.0
django-mssql-backend==2.8.1
django-pyodbc==1.1.3
djangorestframework==3.13.1
djangorestframework-datatables==0.7.0
entrypoints==0.4
executing==0.8.2
greenlet==1.1.2
idna==3.3
install==1.3.5
ipykernel==6.9.1
ipython==8.0.1
jedi==0.18.1
jupyter-client==7.1.2
jupyter-core==4.9.2
matplotlib-inline==0.1.3
mssql-django==1.1.2
mypy-extensions==0.4.3
mysql-connector-python==8.0.28
mysqlclient==2.1.0
nest-asyncio==1.5.4
paramiko==2.10.3
parso==0.8.3
pathspec==0.9.0
petl==1.7.8
pickleshare==0.7.5
pilkit==2.0
Pillow==9.1.1
platformdirs==2.5.1
prompt-toolkit==3.0.28
protobuf==3.20.1
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.11.2
PyNaCl==1.5.0
pyodbc==4.0.32
pyspnego==0.5.0
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.19.2
python-ldap @ file:///C:/Users/jbaynes/dev/hpu_applications/PurplePages/python_ldap-3.4.0-cp310-cp310-win_amd64.whl
pytz==2022.1
pywin32==303
pyzmq==22.3.0
requests==2.27.1
six==1.16.0
smbprotocol==1.9.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.35
sqlparse==0.4.2
stack-data==0.2.0
toml==0.10.2
tomli==2.0.1
tornado==6.1
traitlets==5.1.1
tzdata==2022.1
urllib3==1.26.9
wcwidth==0.2.5
